[fix]머지 수정
